requirements.txt
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
decorator==5.1.0
flasgger==0.9.5
Flask==2.0.2
Flask-JWT-Extended==4.3.1
gunicorn==20.1.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
mistune==0.8.4
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pyrsistent==0.18.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
validators==0.18.2
Werkzeug==2.0.2
zipp==3.6.0
